♻️ [Refactoring]: group hooks/ value objects under hooks/model/
Move hook_definition.rs / name.rs / script_path.rs (and name_test.rs /
script_path_test.rs) into hooks/model/. Introduce hooks/model.rs as the
sub-parent. Declare the inner modules as pub(crate) mod hook_definition;
/ pub(crate) mod name; / pub(crate) mod script_path; — pub(crate) mod is
required so the parent hooks.rs can re-export model::name (rustc rejects
re-exporting a private module with E0365), while pub mod is forbidden
because it would introduce a new external public path
crate::hooks::model::* and violate the DoD.

Lift HookDefinition / CommandHook / HttpHook / StubHook into the
sub-parent's namespace via pub(crate) use self::hook_definition::{...}; so
cross-subgroup references from hooks/converter/ can use
super::super::model::* paths. HookName / resolve_script_path are not
re-exported at the sub-parent because they have no cross-subgroup
consumers (HookName is reached externally via the parent re-export
crate::hooks::name::HookName, and resolve_script_path is only used inside
its own leaf).

Add pub(crate) use model::name; to hooks.rs to keep the external path
crate::hooks::name::HookName (referenced from
component/deployment/hook_deploy.rs:8 and _test.rs:5) unchanged — same
pattern as spec-014's pub(crate) use meta::version;.

Delete pub(crate) mod hook_definition; / pub(crate) mod name; /
pub(crate) mod script_path; from hooks.rs, and move
#[cfg(test)] mod name_test; / #[cfg(test)] mod script_path_test; to the
new sub-parent (test form B).

Rewrite the temporary use crate::hooks::hook_definition::*; references
inside hooks/converter/ to super::super::model::{CommandHook, HookDefinition,
HttpHook, StubHook}; via the model sub-parent's namespace re-export.

Refs #247
