+++ src/hooks/converter/copilot.rs
@@ -5,11 +5,11 @@
 
 use serde_json::Value;
 
+use super::super::model::{CommandHook, HttpHook, StubHook};
 use crate::error::PlmError;
 use crate::hooks::converter::{
     generate_matcher_filter, shell_escape, ConversionWarning, ScriptInfo, SourceFormat, SCRIPTS_DIR,
 };
-use crate::hooks::hook_definition::{CommandHook, HttpHook, StubHook};
 
 pub(crate) use super::super::event::copilot::CopilotEventMap;
 use super::converter::{KeyMap, ScriptGenerator, StructureConverter};
