+++ src/hooks/converter/copilot_test.rs
@@ -2,7 +2,7 @@
 
 use serde_json::json;
 
-use super::super::hook_definition::{CommandHook, HttpHook, StubHook};
+use super::super::model::{CommandHook, HttpHook, StubHook};
 use super::converter::{ConversionWarning, EventMap, KeyMap, ScriptGenerator, StructureConverter};
 use super::copilot::{
     CopilotEventMap, CopilotKeyMap, CopilotScriptGenerator, CopilotStructureConverter,
