+++ src/hooks/model/script_path.rs
